cxr-demo/requirements-cxr.txt
# Offline CXR training + TF.js export (not required to browse the static site).
# See docs/cxr-demo.md
tensorflow>=2.15.0,<2.19.0
tensorflowjs>=4.15.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
